Updated to v1.3.4

+++ ultimate-facebook-comments.php
@@ -3,9 +3,9 @@
  * Plugin Name: Ultimate Facebook Comments
  * Plugin URI: https://iamsayan.github.io/ultimate-facebook-comments/
  * Description: Ultimate Facebook Comments plugin will help you to display Facebook Comments box on your website easily. You can use Facebook Comments on your posts or pages.
- * Version: 1.3.3
+ * Version: 1.3.4
  * Author: Sayan Datta
- * Author URI: https://profiles.wordpress.org/infosatech/
+ * Author URI: https://www.sayandatta.com
  * License: GPLv3
  * Text Domain: ultimate-facebook-comments
  * Domain Path: /languages
@@ -35,7 +35,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'UFC_PLUGIN_VERSION', '1.3.3' );
+define( 'UFC_PLUGIN_VERSION', '1.3.4' );
 
 // Internationalization
 add_action( 'plugins_loaded', 'ufc_plugin_load_textdomain' );
@@ -48,6 +48,28 @@ function ufc_plugin_load_textdomain() {
     load_plugin_textdomain( 'ultimate-facebook-comments', false, dirname( plugin_basename( __FILE__ ) ) . '/languages/' ); 
 }
 
+register_activation_hook( __FILE__, 'ufc_plugin_run_on_activation' );
+register_deactivation_hook( __FILE__, 'ufc_plugin_run_on_deactivation' );
+
+function ufc_plugin_run_on_activation() {
+
+    if ( ! current_user_can( 'activate_plugins' ) ) {
+        return;
+    }
+    set_transient( 'ufc-admin-notice-on-activation', true, 5 );
+}
+
+function ufc_plugin_run_on_deactivation() {
+
+    if ( ! current_user_can( 'activate_plugins' ) ) {
+        return;
+    }
+
+    delete_option( 'ufc_plugin_dismiss_rating_notice' );
+    delete_option( 'ufc_plugin_no_thanks_rating_notice' );
+    delete_option( 'ufc_plugin_installed_time' );
+}
+
 //add admin styles and scripts
 function ufc_custom_admin_styles_scripts( $hook ) {
 
@@ -55,7 +77,7 @@ function ufc_custom_admin_styles_scripts( $hook ) {
     if ( strpos( $current_screen->base, 'ultimate-facebook-comments') !== false ) {
         wp_enqueue_style( 'ufc-admin-style', plugins_url( 'admin/assets/css/admin.min.css', __FILE__ ), array(), UFC_PLUGIN_VERSION );
         wp_enqueue_style( 'ufc-style', plugins_url( 'admin/assets/css/style.min.css', __FILE__ ), array(), UFC_PLUGIN_VERSION );      
-        wp_enqueue_style( 'ufc-selectize-theme-css', plugins_url( 'admin/assets/css/selectize.bootstrap3.css', __FILE__ ), array(), '0.12.6' );       
+        wp_enqueue_style( 'ufc-selectize-theme-css', plugins_url( 'admin/assets/css/selectize.css', __FILE__ ), array(), '0.12.6' );       
         
         wp_enqueue_script( 'ufc-script', plugins_url( 'admin/assets/js/admin.min.js', __FILE__ ), array( 'jquery', 'wp-color-picker' ), UFC_PLUGIN_VERSION, true );
         wp_enqueue_script( 'ufc-selectize-js', plugins_url( 'admin/assets/js/selectize.min.js', __FILE__ ), array(), '0.12.6' );
@@ -72,29 +94,25 @@ function ufc_custom_admin_styles_scripts( $hook ) {
 
 add_action( 'admin_enqueue_scripts', 'ufc_custom_admin_styles_scripts' );
 
-function ufc_first_time_cookie_notice() {
+function ufc_frontend_enqueue_scripts() {
 
     $options = get_option('ufc_plugin_global_options');
+
     if( isset($options['ufc_fb_comment_consent_notice_cb']) && ($options['ufc_fb_comment_consent_notice_cb'] == 1) ) {
         wp_enqueue_script( 'ufc-cookie-js', plugin_dir_url( __FILE__ ) . 'public/js/jquery.cookie.min.js', array ( 'jquery' ), '1.4.1', true );
         wp_enqueue_style( 'ufc-consent', plugins_url( 'public/css/consent.min.css', __FILE__ ), array(), UFC_PLUGIN_VERSION );   
         wp_enqueue_script( 'ufc-consent-js', plugin_dir_url( __FILE__ ) . 'public/js/consent.min.js', array ( 'jquery' ), UFC_PLUGIN_VERSION, true );
     }
-}
-
-add_action( 'wp_enqueue_scripts', 'ufc_first_time_cookie_notice' );
-
-function ufc_frontend_enqueue_scripts() {
     /**
      * frontend ajax requests.
      */
     wp_enqueue_script( 'ufc-frontend-script', plugins_url( 'public/js/frontend.min.js', __FILE__ ), array( 'jquery' ), UFC_PLUGIN_VERSION, true );
     wp_localize_script( 'ufc-frontend-script', 'ufc_frontend_ajax_data',
         array( 
-            'ajaxurl' => admin_url( 'admin-ajax.php' ),
+            'ajaxurl'   => admin_url( 'admin-ajax.php' ),
             'permalink' => get_permalink(),
-            'title' => get_the_title(),
-            'postid' => get_the_ID()
+            'title'     => get_the_title(),
+            'postid'    => get_the_ID()
         )
     );
 }
@@ -113,40 +131,32 @@ function ufc_ajax_save_admin_scripts() {
 
 add_action( 'admin_init', 'ufc_ajax_save_admin_scripts' );
 
-register_activation_hook( __FILE__, 'ufc_plugin_run_on_activation' );
-register_deactivation_hook( __FILE__, 'ufc_plugin_run_on_deactivation' );
+function ufc_plugin_all_notices() { 
+    $options = get_option('ufc_plugin_global_options');
 
-function ufc_plugin_run_on_activation() {
-
-    if ( ! current_user_can( 'activate_plugins' ) ) {
-        return;
-    }
-    set_transient( 'ufc-admin-notice-on-activation', true, 20 );
-}
-
-function ufc_plugin_run_on_deactivation() {
-
-    if ( ! current_user_can( 'activate_plugins' ) ) {
-        return;
-    }
-
-    delete_option( 'ufc_plugin_dismiss_rating_notice' );
-    delete_option( 'ufc_plugin_no_thanks_rating_notice' );
-    delete_option( 'ufc_plugin_installed_time' );
-}
-
-function ufc_plugin_install_notice() { 
     /* Check transient, if available display notice */
-    if( get_transient( 'ufc-admin-notice-on-activation' ) ) { ?>
+    if( get_transient( 'ufc-admin-notice-on-activation' ) && empty( $options['ufc_facebook_comments_app_id'] ) ) { ?>
         <div class="notice notice-success is-dismissible">
             <p><strong><?php printf( __( 'Thanks for installing %1$s v%2$s plugin. Click <a href="%3$s">here</a> to configure plugin settings.', 'ultimate-facebook-comments' ), 'Ultimate Facebook Comments', UFC_PLUGIN_VERSION, admin_url( 'admin.php?page=ultimate-facebook-comments' ) ); ?></strong></p>
         </div> <?php
         /* Delete transient, only display this notice once. */
         delete_transient( 'ufc-admin-notice-on-activation' );
     }
+
+    if( empty( $options['ufc_facebook_comments_app_id'] ) && empty( $options['ufc_facebook_comments_app_secret'] ) ) { ?>
+        <div id="appid-notice" class="notice notice-error">
+            <p><?php _e( 'Please create and enter your Facebook App ID for Ultimate Facebook Comments Plugin to work properly. If you have already created an app, then put <strong>APP ID</strong> and <strong>APP Secret</strong> in plugin settings.', 'ultimate-facebook-comments' ); ?></p>
+            <p><a href="https://developers.facebook.com/apps" target="_blank" class="button button-secondary"><?php _e( 'Create New Facebook App ID', 'ultimate-facebook-comments' ); ?></a></p>
+        </div><?php
+    } elseif( !empty( $options['ufc_facebook_comments_app_id'] ) && empty( $options['ufc_facebook_comments_app_secret'] ) ) { ?>
+        <div id="appid-notice" class="notice notice-error">
+            <p><?php _e( 'Please enter your Facebook App Secret for Ultimate Facebook Comments Plugin to work properly. Go to Facebook for Developers and copy App Secret and paste it to plugin settings.', 'ultimate-facebook-comments' ); ?></p>
+            <p><a href="https://developers.facebook.com/apps/<?php echo $options['ufc_facebook_comments_app_id']; ?>/settings/basic/" target="_blank" class="button button-secondary"><?php _e( 'View App Secret', 'ultimate-facebook-comments' ); ?></a></p>
+        </div><?php
+    }
 }
 
-add_action( 'admin_notices', 'ufc_plugin_install_notice' ); 
+add_action( 'admin_notices', 'ufc_plugin_all_notices' ); 
 
 $options = get_option('ufc_plugin_global_options');
 
@@ -154,36 +164,12 @@ require_once plugin_dir_path( __FILE__ ) . 'admin/loader.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/notice.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/ajax.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/disable-native.php';
+require_once plugin_dir_path( __FILE__ ) . 'admin/admin-bar.php';
+require_once plugin_dir_path( __FILE__ ) . 'admin/dashboard-edit-screen.php';
+require_once plugin_dir_path( __FILE__ ) . 'admin/dashboard-column.php';
+require_once plugin_dir_path( __FILE__ ) . 'public/comments-loader.php';
 require_once plugin_dir_path( __FILE__ ) . 'public/shortcode.php';
 require_once plugin_dir_path( __FILE__ ) . 'public/template-tags.php';
-
-if( isset($options['ufc_enable_fb_comment_cb']) && ($options['ufc_enable_fb_comment_cb'] == 1) ) {
-    require_once plugin_dir_path( __FILE__ ) . 'admin/admin-bar.php';
-    require_once plugin_dir_path( __FILE__ ) . 'admin/dashboard-edit-screen.php';
-    require_once plugin_dir_path( __FILE__ ) . 'admin/dashboard-column.php';
-    require_once plugin_dir_path( __FILE__ ) . 'public/comments-loader.php';
-}
-
-add_action( 'admin_notices', 'ufc_no_app_id_notice' );
-
-function ufc_no_app_id_notice() {
-
-    $options = get_option('ufc_plugin_global_options');
-
-    if ( ! current_user_can( 'manage_options' ) ) {
-        return;
-    }
-
-    if( !empty( $options['ufc_facebook_comments_app_id'] ) ) return;
-
-    $current_screen = get_current_screen();
-    if ( strpos( $current_screen->base, 'ultimate-facebook-comments') !== false ) { ?>
-        <div id="appid-notice" class="notice notice-error">
-            <p style="line-height: 2.5;"><?php _e( 'Please create and enter your Facebook App ID for Ultimate Facebook Comments Plugin to work properly.', 'ultimate-facebook-comments' ); ?>&nbsp;&nbsp;<a href="https://developers.facebook.com/apps" target="_blank" class="button button-secondary"><?php _e( 'Create App ID', 'ultimate-facebook-comments' ); ?></a></p>
-        </div>
-    <?php
-    }
-}
 
 // add action links
 add_filter( 'plugin_action_links_' . plugin_basename(__FILE__), 'ufc_add_action_links', 10, 2 );
